+++ apps/preview/nuxt/pages/showcases/Gallery/GalleryVertical.vue
@@ -29,7 +29,7 @@
         type="button"
         :aria-label="alt"
         :aria-current="activeIndex === index"
-        :class="`md:w-[78px] md:h-auto relative shrink-0 pb-1 mx-4 -mb-2 border-b-4 snap-start cursor-pointer focus-visible:outline focus-visible:outline-offset transition-colors flex-grow md:grow-0  ${
+        :class="`md:w-[78px] md:h-auto relative shrink-0 pb-1 mx-4 -mb-2 border-b-4 snap-start cursor-pointer focus-visible:outline focus-visible:outline-offset transition-colors grow md:grow-0  ${
           activeIndex === index ? 'border-primary-700' : 'border-transparent'
         }`"
         @mouseover="activeIndex = index"